requirements.txt
mcp
httpx
xmltodict
pydantic
python-dotenv
fastapi
uvicorn[standard]
google-cloud-secret-manager
Authlib
